requirements.txt
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
idna==2.8
PyYAML==5.1.1
soupsieve==1.9.1
urllib3==1.25.3
